blog/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pillow>=10.3.0
python-slugify>=8.0.1